Sync code in local with cluster

+++ src/sampling.py
@@ -33,7 +33,7 @@ def mnist_noniid(dataset, num_users):
     # 60,000 training imgs -->  200 imgs/shard X 300 shards
     num_shards, num_imgs = 200, 300
     idx_shard = [i for i in range(num_shards)]
-    dict_users = {i: np.array([]) for i in range(num_users)}
+    dict_users = {i: np.array([], dtype=np.int64) for i in range(num_users)}
     idxs = np.arange(num_shards*num_imgs)
     labels = dataset.train_labels.numpy()
 
@@ -48,7 +48,7 @@ def mnist_noniid(dataset, num_users):
         idx_shard = list(set(idx_shard) - rand_set)
         for rand in rand_set:
             dict_users[i] = np.concatenate(
-                (dict_users[i], idxs[rand*num_imgs:(rand+1)*num_imgs]), axis=0)
+                (dict_users[i], idxs[rand*num_imgs:(rand+1)*num_imgs]), axis=0, dtype=np.int)
     return dict_users
 
 
@@ -64,7 +64,7 @@ def mnist_noniid_unequal(dataset, num_users):
     # 60,000 training imgs --> 50 imgs/shard X 1200 shards
     num_shards, num_imgs = 1200, 50
     idx_shard = [i for i in range(num_shards)]
-    dict_users = {i: np.array([]) for i in range(num_users)}
+    dict_users = {i: np.array([], dtype=np.int64) for i in range(num_users)}
     idxs = np.arange(num_shards*num_imgs)
     labels = dataset.train_labels.numpy()
 
@@ -167,7 +167,7 @@ def cifar_noniid(dataset, num_users):
     """
     num_shards, num_imgs = 200, 250
     idx_shard = [i for i in range(num_shards)]
-    dict_users = {i: np.array([]) for i in range(num_users)}
+    dict_users = {i: np.array([], dtype=np.int64) for i in range(num_users)}
     idxs = np.arange(num_shards*num_imgs)
     # labels = dataset.train_labels.numpy()
     labels = np.array(dataset.train_labels)
